ElectionServer/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
sqlparse==0.5.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
yarl==1.14.0
psycopg2